[refactor] BottomSheet, Header, SocialIcon components
