+++ src/BottomSheet/index.tsx
@@ -1,0 +1,6 @@
+import { BottomSheet, BottomSheetProps } from './BottomSheet';
+import { withTheme } from '../config';
+
+export { BottomSheet };
+export type { BottomSheetProps };
+export default withTheme(BottomSheet, 'BottomSheet');
