+++ src/Header/Header.tsx
@@ -5,7 +5,6 @@ import {
   StyleSheet,
   View,
   ImageBackground,
-  TextStyle,
   StyleProp,
   TextProps,
   ViewProps,
@@ -14,56 +13,11 @@ import {
   ImageSourcePropType,
   ImageStyle,
   ViewStyle,
-  FlexAlignType,
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { withTheme } from '../config';
-import { renderNode, RneFunctionComponent } from '../helpers';
-
-import Text from '../Text';
-import Icon, { IconObject } from '../icons/Icon';
-
-type Placement = 'left' | 'center' | 'right';
-
-const ALIGN_STYLE: Record<Placement, FlexAlignType> = {
-  left: 'flex-start',
-  right: 'flex-end',
-  center: 'center',
-};
-
-type HeaderChildrenProps = {
-  placement: Placement;
-  style: StyleProp<ViewStyle>;
-  children: any;
-};
-
-const Children = ({ style, placement, children }: HeaderChildrenProps) => (
-  <View
-    style={StyleSheet.flatten([{ alignItems: ALIGN_STYLE[placement] }, style])}
-  >
-    {children == null || children === false
-      ? null
-      : children.text
-      ? renderNode(Text, children.text, { numberOfLines: 1, ...children })
-      : children.icon
-      ? renderNode(Icon, {
-          ...children,
-          name: children.icon,
-          containerStyle: StyleSheet.flatten([
-            { alignItems: ALIGN_STYLE[placement] },
-            children.containerStyle,
-          ]),
-        })
-      : renderNode(Text, children)}
-  </View>
-);
-
-interface HeaderIcon extends IconObject {
-  icon?: string;
-  text?: string;
-  color?: string;
-  style?: StyleProp<TextStyle>;
-}
+import { RneFunctionComponent } from '../helpers';
+import { Children } from './components/HeaderChildren';
+import { HeaderIcon } from './components/HeaderIcon';
 
 type HeaderSubComponent = React.ReactElement<{}> | TextProps | HeaderIcon;
 
@@ -87,39 +41,36 @@ export type HeaderProps = ViewProps & {
   elevated?: boolean;
 };
 
-const Header: RneFunctionComponent<HeaderProps> = (props) => {
+export const Header: RneFunctionComponent<HeaderProps> = ({
+  statusBarProps,
+  leftComponent,
+  centerComponent,
+  rightComponent,
+  leftContainerStyle,
+  centerContainerStyle,
+  rightContainerStyle,
+  backgroundColor,
+  backgroundImage,
+  backgroundImageStyle,
+  containerStyle,
+  placement = 'center',
+  barStyle,
+  children = [],
+  linearGradientProps,
+  ViewComponent = linearGradientProps || !backgroundImage
+    ? View
+    : ImageBackground,
+  theme,
+  elevated,
+  ...attributes
+}) => {
   React.useEffect(() => {
-    const { linearGradientProps, ViewComponent } = props;
     if (linearGradientProps && !ViewComponent) {
       console.error(
         "You need to pass a ViewComponent to use linearGradientProps !\nExample: ViewComponent={require('react-native-linear-gradient')}"
       );
     }
   });
-
-  const {
-    statusBarProps,
-    leftComponent,
-    centerComponent,
-    rightComponent,
-    leftContainerStyle,
-    centerContainerStyle,
-    rightContainerStyle,
-    backgroundColor,
-    backgroundImage,
-    backgroundImageStyle,
-    containerStyle,
-    placement = 'center',
-    barStyle,
-    children = [],
-    linearGradientProps,
-    ViewComponent = linearGradientProps || !backgroundImage
-      ? View
-      : ImageBackground,
-    theme,
-    elevated,
-    ...attributes
-  } = props;
 
   return (
     <>
@@ -220,6 +171,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export { Header };
-
-export default withTheme<HeaderProps, {}>(Header, 'Header');
+Header.displayName = 'Header';
