+++ src/Header/components/HeaderIcon.tsx
@@ -1,0 +1,9 @@
+import { IconObject } from '../../icons/Icon';
+import { StyleProp, TextStyle } from 'react-native';
+
+export type HeaderIcon = IconObject & {
+  icon?: string;
+  text?: string;
+  color?: string;
+  style?: StyleProp<TextStyle>;
+};
