+++ src/Header/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { Header, HeaderProps } from './Header';
+
+export { Header };
+export type { HeaderProps };
+export default withTheme<HeaderProps, {}>(Header, 'Header');
