+++ src/SocialIcon/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { SocialIcon, SocialIconProps, SocialMediaType } from './SocialIcon';
+
+export { SocialIcon };
+export type { SocialIconProps, SocialMediaType };
+export default withTheme(SocialIcon, 'SocialIcon');
