+++ src/config/theme.tsx
@@ -3,13 +3,13 @@ import { TapRatingProps, SwipeRatingProps } from 'react-native-ratings';
 import { AccessoryProps } from '../avatar/Accessory';
 import { AvatarProps } from '../avatar/Avatar';
 import { BadgeProps } from '../badge/Badge';
-import { BottomSheetProps } from '../bottomSheet/BottomSheet';
+import { BottomSheetProps } from '../BottomSheet';
 import { ButtonProps } from '../buttons/Button';
 import { ButtonGroupProps } from '../buttons/ButtonGroup';
 import { CardProps } from '../card/Card';
 import { CheckBoxProps } from '../checkbox/CheckBox';
 import { DividerProps } from '../Divider';
-import { HeaderProps } from '../header/Header';
+import { HeaderProps } from '../Header';
 import { IconProps } from '../icons/Icon';
 import { ImageProps } from '../image/Image';
 import { InputProps } from '../input/Input';
@@ -18,7 +18,7 @@ import { OverlayProps } from '../overlay/Overlay';
 import { PricingCardProps } from '../pricing/PricingCard';
 import { SearchBarProps } from '../searchbar/SearchBar';
 import { SliderProps } from '../slider/Slider';
-import { SocialIconProps } from '../social/SocialIcon';
+import { SocialIconProps } from '../SocialIcon';
 import { TextProps } from '../Text';
 import { TileProps } from '../tile/Tile';
 import { TooltipProps } from '../tooltip/Tooltip';
